Setting Up Git Workflow and adding Hydrapaper Functionality
First time using Github desktop, so I finally got the project synced with the local files I have to better auto-update the github page.

While getting that working, I added the missing functionality for hydrapaper, so it will now work with two monitors.

I also added continually checking of the config file to see if it was updated, so changed can be made during the runtime. As a result, the program will no longer end when bad config settings are found, and will instead not display anything while it waits for the config file to update again. This is to hopefully make it easier to work with when autostarted, so you don't have to find and kill the process to make changes, or find and start the file if the config file didn't work as anticipated.